% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_lognorm_fit}
\alias{add_lognorm_fit}
\title{add lognormal fit}
\usage{
add_lognorm_fit(df, c, h, w, b)
}
\arguments{
\item{c}{mean}

\item{h}{height}

\item{w}{width}

\item{b}{asymmetry}

\item{lambda}{wavelength}
}
\description{
add lognormal fit
}

