% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_normal_fit}
\alias{add_normal_fit}
\title{add normal fit}
\usage{
add_normal_fit(df, c, h, w)
}
\arguments{
\item{c}{mean}

\item{h}{height}

\item{w}{width}

\item{lambda}{wavelength}
}
\description{
add normal fit
}

