% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{import data from csv file}
\usage{
import_data(path, filename, skip = 2, nlines = 121, scans = 15)
}
\arguments{
\item{path}{path to data source}

\item{filename}{filename}

\item{skip}{remove header lines}

\item{nlines}{number of scan points}

\item{scans}{number of scans}
}
\value{
data.frame containing data labled R1..RN where N is the number of scans
}
\description{
This function imports the data from spectroscopy measurements
}

