% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{melt_data}
\alias{melt_data}
\title{melts data in preparation for ggplot2 plots}
\usage{
melt_data(df)
}
\arguments{
\item{df}{dataframe of spectroscopy data imported using caryscan package}
}
\description{
melts data in preparation for ggplot2 plots
}

