% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all}
\alias{plot_all}
\title{plot all the data}
\usage{
plot_all(df, scans = 15)
}
\arguments{
\item{df}{dataframe of spectroscopy data imported using caryscan package}

\item{scans}{number of spectra}
}
\value{
plot
}
\description{
This function prepares a plot of all the spectroscopy data
}

