% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_all_gg}
\alias{plot_all_gg}
\title{ggplot all the data}
\usage{
plot_all_gg(df, mean_name = "fluo_mean")
}
\arguments{
\item{df}{dataframe of spectroscopy data imported using caryscan package}

\item{mean_name}{the name of the column containing the averaged data}
}
\value{
ggplot2 plot object
}
\description{
This function prepares a ggplot2 plot of all the spectroscopy data
}

