% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fits_gg}
\alias{plot_fits_gg}
\title{ggplot the fits data}
\usage{
plot_fits_gg(dat, fit_data, limits = c(275, 400))
}
\arguments{
\item{fit_data}{dataframe of fits for each peak}

\item{limits}{vector of wavelength range for plotting}

\item{df}{dataframe of spectroscopy data imported using caryscan package}
}
\value{
ggplot2 plot object
}
\description{
This function prepares a ggplot2 plot of the fits
}

