% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mean}
\alias{plot_mean}
\title{plot the averaged data}
\usage{
plot_mean(df, mean_name)
}
\arguments{
\item{df}{dataframe of spectroscopy data imported using caryscan package}

\item{mean_name}{the name of the column containing the averaged data}
}
\value{
plot
}
\description{
This function prepares a plot of the averaged spectroscopy data
}

