% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_pretty_gg}
\alias{plot_pretty_gg}
\title{ggplot the main and residuals plots together}
\usage{
plot_pretty_gg(main_plot, res_plot)
}
\arguments{
\item{ggplot}{object for main plot}

\item{ggplot}{object for residuals plot}
}
\value{
ggplot2 plot object
}
\description{
This function prepares a ggplot2 plot of the residuals from the fits
}

