% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_residuals_gg}
\alias{plot_residuals_gg}
\title{ggplot the residuals data}
\usage{
plot_residuals_gg(fit_data, colname = "residuals")
}
\arguments{
\item{fit_data}{dataframe of fits for each peak}

\item{df}{dataframe of spectroscopy data imported using caryscan package}

\item{limits}{vector of wavelength range for plotting}
}
\value{
ggplot2 plot object
}
\description{
This function prepares a ggplot2 plot of the residuals from the fits
}

