% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_fits}
\alias{sum_fits}
\title{sum fit columns}
\usage{
sum_fits(fit_data, cols)
}
\arguments{
\item{fit_data}{dataframe containing fit data}

\item{cols}{names of columns containing fit data #not yet implemented}
}
\description{
sum fit columns
}

