% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_mean}
\alias{write_mean}
\title{write average data to file}
\usage{
write_mean(df, path)
}
\arguments{
\item{df}{dataframe of spectroscopy data imported using caryscan package}

\item{path}{to filename}
}
\description{
write average data to file
}

